# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AqariTalk API — AI-guided real estate workflow platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/otp/request:
    post:
      operationId: requestOtp
      tags: [auth]
      summary: Request OTP code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OtpRequest"
      responses:
        "200":
          description: OTP sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OtpRequestResponse"
        "429":
          description: Rate limited
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/otp/verify:
    post:
      operationId: verifyOtp
      tags: [auth]
      summary: Verify OTP and authenticate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OtpVerify"
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Invalid or expired code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateMe
      tags: [auth]
      summary: Update current user profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  # ── Properties ────────────────────────────────────────
...
  /admin/users:
    get:
      operationId: adminListUsers
      tags: [admin]
      summary: List all users
      security:
        - bearerAuth: []
      parameters:
        - name: role
          in: query
          schema:
            type: ["string", "null"]
        - name: status
          in: query
          schema:
            type: ["string", "null"]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
  /admin/users/{id}:
    put:
      operationId: adminUpdateUser
      tags: [admin]
      summary: Update user (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /admin/properties:
    get:
      operationId: adminListProperties
      tags: [admin]
      summary: List all properties (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: ["string", "null"]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Properties
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PropertyList"
  /admin/properties/{id}/status:
    put:
      operationId: adminUpdatePropertyStatus
      tags: [admin]
      summary: Approve or reject a property listing
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminPropertyStatusUpdate"
      responses:
        "200":
          description: Updated property
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Property"
  /admin/stats:
    get:
      operationId: getAdminStats
      tags: [admin]
      summary: Platform statistics
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStats"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string

    # ── Auth schemas ──────────────────────────────────
    OtpRequest:
      type: object
      required: [phone]
      properties:
        phone:
          type: string
          minLength: 7
          maxLength: 20
    OtpRequestResponse:
      type: object
      required: [sessionId, expiresAt]
      properties:
        sessionId:
          type: string
        expiresAt:
          type: string
          format: date-time
    OtpVerify:
      type: object
      required: [phone, code]
      properties:
        phone:
          type: string
        code:
          type: string
          minLength: 6
          maxLength: 6
    AuthResponse:
      type: object
      required: [token, user, isNewUser]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
        isNewUser:
          type: boolean

    # ── User schemas ──────────────────────────────────
    User:
      type: object
      required: [id, phone, role, market, language, verificationStatus, status, autoSendVoice, createdAt, updatedAt]
      properties:
        id:
          type: string
        phone:
          type: string
        name:
          type: ["string", "null"]
        role:
          type: string
          enum: [buyer, seller, broker, admin]
        market:
          type: string
          enum: [JO, SA, AE, EG, KW, QA, BH, OM, MA, LB, IQ]
        language:
          type: string
          enum: [ar, en]
        verificationStatus:
          type: string
          enum: [unverified, verified]
        status:
          type: string
          enum: [active, suspended]
        autoSendVoice:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    UserUpdate:
      type: object
      properties:
        name:
          type: string
        role:
          type: string
          enum: [buyer, seller, broker]
        market:
          type: string
          enum: [JO, SA, AE, EG, KW, QA, BH, OM, MA, LB, IQ]
        language:
          type: string
          enum: [ar, en]
        autoSendVoice:
          type: boolean

    # ── Property schemas ──────────────────────────────
    Property:
      type: object
      required: [id, createdBy, propertyType, transactionMode, listingDirection, status, brokerListing, createdAt, updatedAt]
      properties:
        id:
          type: string
        createdBy:
          type: string
        listingName:
          type: ["string", "null"]
        listingDirection:
          type: string
          enum: [offering, seeking]
        propertyType:
          type: string
        transactionMode:
          type: string
          enum: [sale, rent, lease]
        rentalPeriod:
          type: ["string", "null"]
        price:
          type: ["number", "null"]
        priceCurrency:
          type: string
        priceNegotiable:
          type: boolean
        country:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        district:
          type: ["string", "null"]
        street:
          type: ["string", "null"]
        addressFull:
          type: ["string", "null"]
        latitude:
          type: ["number", "null"]
        longitude:
          type: ["number", "null"]
        locationAccuracy:
          type: ["string", "null"]
        areaSqm:
          type: ["number", "null"]
        rooms:
          type: ["integer", "null"]
        bathrooms:
          type: ["integer", "null"]
        floorNumber:
          type: ["integer", "null"]
        furnishedStatus:
          type: ["string", "null"]
        parking:
          type: ["boolean", "null"]
        hasElevator:
          type: ["boolean", "null"]
        condition:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [draft, pending_review, active, sold, rented, expired, rejected, deleted]
        verified:
          type: boolean
        aiMissingFields:
          type: ["array", "null"]
          items:
            type: string
        contactPreference:
          type: ["string", "null"]
        brokerListing:
          type: boolean
        images:
          type: array
          items:
            $ref: "#/components/schemas/PropertyImage"
        distance:
          type: ["number", "null"]
          description: Distance in km from search center (only present in search results)
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    PropertyList:
      type: object
      required: [items, total, page, limit]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Property"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    PropertyImage:
      type: object
      required: [id, propertyId, path, isVoiceNote, createdAt]
      properties:
        id:
          type: string
        propertyId:
          type: string
        path:
          type: string
        gpsLat:
          type: ["number", "null"]
        gpsLng:
          type: ["number", "null"]
        sizeBytes:
          type: ["integer", "null"]
        isVoiceNote:
          type: boolean
        createdAt:
          type: string
          format: date-time

    # ── Conversation schemas ──────────────────────────
...
    AdminUserUpdate:
      type: object
      properties:
        role:
          type: string
          enum: [buyer, seller, broker, admin]
        status:
          type: string
          enum: [active, suspended]
        verificationStatus:
          type: string
          enum: [unverified, verified]
    AdminPropertyStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [active, rejected, deleted]
        rejectionReason:
          type: string
    AdminStats:
      type: object
      required: [totalUsers, activeListings, contactReleasesThisMonth, pendingReview]
      properties:
        totalUsers:
          type: integer
        activeListings:
          type: integer
        contactReleasesThisMonth:
          type: integer
        pendingReview:
          type: integer
        usersByRole:
          type: object
          properties:
            buyer:
              type: integer
            seller:
              type: integer
            broker:
              type: integer