# lib/api-zod/src/generated/types/otpRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AqariTalk API — AI-guided real estate workflow platform
 * OpenAPI spec version: 0.1.0
 */

export interface OtpRequest {
  /**
     * @minLength 7
     * @maxLength 20
     */
  phone: string;
}