# lib/api-zod/src/generated/types/property.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AqariTalk API — AI-guided real estate workflow platform
 * OpenAPI spec version: 0.1.0
 */
import type { PropertyImage } from './propertyImage';
import type { PropertyListingDirection } from './propertyListingDirection';
import type { PropertyStatus } from './propertyStatus';
import type { PropertyTransactionMode } from './propertyTransactionMode';

export interface Property {
  id: string;
  createdBy: string;
  /** @nullable */
  listingName?: string | null;
  listingDirection: PropertyListingDirection;
  propertyType: string;
  transactionMode: PropertyTransactionMode;
  /** @nullable */
  rentalPeriod?: string | null;
  /** @nullable */
  price?: number | null;
  priceCurrency?: string;
  priceNegotiable?: boolean;
  /** @nullable */
  country?: string | null;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  district?: string | null;
  /** @nullable */
  street?: string | null;
  /** @nullable */
  addressFull?: string | null;
  /** @nullable */
  latitude?: number | null;
  /** @nullable */
  longitude?: number | null;
  /** @nullable */
  locationAccuracy?: string | null;
  /** @nullable */
  areaSqm?: number | null;
  /** @nullable */
  rooms?: number | null;
  /** @nullable */
  bathrooms?: number | null;
  /** @nullable */
  floorNumber?: number | null;
  /** @nullable */
  furnishedStatus?: string | null;
  /** @nullable */
  parking?: boolean | null;
  /** @nullable */
  hasElevator?: boolean | null;
  /** @nullable */
  condition?: string | null;
  /** @nullable */
  description?: string | null;
  status: PropertyStatus;
  verified?: boolean;
  /** @nullable */
  aiMissingFields?: string[] | null;
  /** @nullable */
  contactPreference?: string | null;
  brokerListing: boolean;
  images?: PropertyImage[];
  /**
     * Distance in km from search center (only present in search results)
     * @nullable
     */
  distance?: number | null;
  createdAt: Date;
  updatedAt: Date;
}